// types.ts
export type ResponseDataItem = {
    attributes: {
        altitude: string;
        city: string;
        country: string;
        iata: string;
        icao: string;
        latitude: string;
        longitude: string;
        name: string;
        timezone: string;
    };
    id: string;
    type: string;
};